fix: 반환형 integer -> long 으로 변경

+++ src/main/java/com/example/wms/inbound/adapter/in/dto/request/InboundPutAwayReqDto.java
@@ -17,6 +17,6 @@ public class InboundPutAwayReqDto {
     private Long productId;
 
     @Schema(description = "품목 별 lot 단위 개수", example = "3")
-    private Integer lotCount;
+    private Long lotCount;
 
 }
